working on error mannagment and adding more tests

+++ Test/Test/compatebility_system_test_hooks.hpp
@@ -1,13 +1,15 @@
 #ifndef COMPATEBILITY_SYSTEM_TEST_HOOKS_HPP
 #define COMPATEBILITY_SYSTEM_TEST_HOOKS_HPP
 
+#include <cstdint>
+
 void    cmp_set_force_cross_device_move(int force_cross_device_move);
 
 void    cmp_set_force_cpu_count_success(unsigned int cpu_count);
 void    cmp_set_force_cpu_count_failure(int errno_value);
 void    cmp_clear_force_cpu_count_result(void);
 
-void    cmp_set_force_total_memory_success(unsigned long long memory_size);
+void    cmp_set_force_total_memory_success(uint64_t memory_size);
 void    cmp_set_force_total_memory_failure(int errno_value);
 #if defined(_WIN32) || defined(_WIN64)
 void    cmp_set_force_total_memory_windows_failure(unsigned long last_error);
